Change behaviour; new icon; change UUID
Change beheviour:

Now by default it overrides the system brightness slider button to
toggle auto brightness.

You have an option in preferences to add a quick settings toggle (like
older versions) as well.

Now on unsupported computers the extension will throw an error and will
not activate (in the mind of that if it activates, then it should work).

New icon: new symbolic icon from Icon Library of libadwaita

Change UUID: the old domain name has expired and hence the UUID is
changed.